Implemented AP mode setting. Consolidated field validators.
