+++ endaqconfig/base.py
@@ -35,7 +35,7 @@ import wx.lib.filebrowsebutton as FB
 import wx.lib.scrolledpanel as SP
 
 from .common import getUtcOffset, isCompiled
-from .widgets.shared import DateTimeCtrl, wx_DateTime_FromTimeT
+from .widgets.shared import DateTimeCtrl, wx_DateTime_FromTimeT, TextValidator
 
 # ===============================================================================
 # --- Utility functions
@@ -81,95 +81,9 @@ def getConfigId(el: Element) -> Optional[int]:
 #
 # ===============================================================================
 
-def getClipboardText():
-    """ Retrieve text from the clipboard.
-    """
-    if not wx.TheClipboard.IsOpened():
-        wx.TheClipboard.Open()
-
-    obj = wx.TextDataObject()
-    if wx.TheClipboard.GetData(obj):
-        return obj.GetText()
-
-    return ""
-
-
 # ===============================================================================
 #
 # ===============================================================================
-
-class TextValidator(wx.Validator):
-    """ Generic Validator for TextField and ASCIIField text widgets.
-    """
-
-    VALID_KEYS = (wx.WXK_LEFT, wx.WXK_UP, wx.WXK_RIGHT, wx.WXK_DOWN,
-                  wx.WXK_HOME, wx.WXK_END, wx.WXK_PAGEUP, wx.WXK_PAGEDOWN,
-                  wx.WXK_INSERT, wx.WXK_DELETE)
-
-
-    def __init__(self, validator, maxLen=None):
-        """ Instantiate a text field validator.
-
-            :keyword validator: A function that validates the string.
-        """
-        self.maxLen = maxLen
-        self.isValid = validator
-        wx.Validator.__init__(self)
-        self.Bind(wx.EVT_CHAR, self.OnChar)
-        self.Bind(wx.EVT_TEXT_PASTE, self.OnPaste)
-
-
-    def Clone(self):
-        """ Required in wx.PyValidator subclasses. """
-        return TextValidator(self.isValid, self.maxLen)
-
-
-    def TransferToWindow(self):
-        """ Required in wx.PyValidator subclasses. """
-        return True
-
-
-    def TransferFromWindow(self):
-        """ Required in wx.PyValidator subclasses. """
-        return True
-
-
-    def Validate(self, win):
-        txt = self.GetWindow().GetValue()
-        return self.isValid(txt)
-
-
-    def OnChar(self, evt):
-        """ Validate a character that has been typed.
-        """
-        key = evt.GetKeyCode()
-
-        if key < wx.WXK_SPACE or key in self.VALID_KEYS:
-            evt.Skip()
-            return
-
-        val = self.GetWindow().GetValue()
-
-        if self.isValid(chr(key)) and len(val) < self.maxLen:
-            evt.Skip()
-            return
-
-        if not wx.Validator.IsSilent():
-            wx.Bell()
-
-        return
-
-
-    def OnPaste(self, evt):
-        """ Validate text pasted into the field.
-        """
-        txt = getClipboardText()
-        current = self.GetWindow().GetValue()
-        if self.isValid(current + txt):
-            evt.Skip()
-        elif not wx.Validator.IsSilent():
-            wx.Bell()
-
 
 # ===============================================================================
 #
